feat(SPV-1493) : Update conflicts
chore(paymail_query) : update paymail query utility with new js client format
